+++ frontend/src/dashboard.css
@@ -1,3 +1,7 @@
 .sidebar-item {
   @apply flex h-14 text-xl font-light text-gray-600 cursor-pointer hover:text-primary
 }
+
+.sidebar-item.selected {
+  @apply text-primary
+}
